create separate file with downloading functions
