+++ bin/link_MAG_to_primary_metagenome_assembly.py
@@ -5,22 +5,20 @@ import argparse
 import csv
 import json
 import gzip
-from ftplib import FTP
+import hashlib
 import logging
 import os
 import re
 import shutil
 import urllib.parse
 
-import boto3
-from botocore import UNSIGNED
-from botocore.config import Config
+from Bio import SeqIO
 from botocore.exceptions import ClientError, ParamValidationError
 import requests
 import xmltodict
 from tqdm import tqdm
 from retry import retry
-from mag_assembly_checksum_compare import compute_hashes,get_fasta_url
+from download_fasta_utils import download_from_ENA_FIRE, download_from_ENA_API, download_from_ENA_FTP
 
 # TODO add docs for functions and Type Annotations
 # TODO look for primary assemblies even if bin sample is bio sample? 
@@ -279,109 +277,6 @@ def handle_fasta_processing(accession, download_folder):
         return None
 
 
-@retry(tries=5, delay=15, backoff=1.5) 
-def download_from_ENA_FIRE(accession: str, analysis_ftp_field: str, outpath: str) -> str:
-    url = get_fasta_url(accession, analysis_ftp_field=analysis_ftp_field)
-    if not url:
-        logging.debug(f"{accession} URL is empty for accession, ftp field: {analysis_ftp_field}")
-        return None
-        # raise ValueError(f"URL is empty, ftp field: {analysis_ftp_field}")
-    logging.debug(f"Download {accession} from ENA FIRE using URL {url}")
-
-    fire_endpoint = "http://hl.fire.sdo.ebi.ac.uk"
-    fire_ena_bucket = "era-public"
-    fire_path = url.replace("ftp.sra.ebi.ac.uk/vol1/", "")
-    s3 = boto3.client("s3", endpoint_url=fire_endpoint, config=Config(signature_version=UNSIGNED))
-    s3.download_file(fire_ena_bucket, fire_path, outpath)
-    # 20 bytes is a size of an empty fa.gz
-    if os.path.exists(outpath) and os.path.getsize(outpath) > 20:
-        logging.debug(f"Successful. File saved to {outpath}")
-        return outpath
-    logging.debug(f"Downloaded file {outpath} has zero size. Removing the file.")
-    os.remove(outpath)
-    return None
-    # raise ValueError(f"Downloaded file {outpath} has zero size")
-
-
-@retry(tries=7, delay=15, backoff=2) 
-def download_from_ENA_API(accession: str, outpath: str) -> str:
-    api_endpoint = f"https://www.ebi.ac.uk/ena/browser/api/fasta/{accession}"
-    logging.debug(f"Download {accession} from ENA API using URL {api_endpoint}")
-    query = {
-        'download': 'true',
-        'gzip': 'true'
-    }
-    response = requests.get(api_endpoint, params=urllib.parse.urlencode(query))
-    response.raise_for_status()
-    
-    with open(outpath, 'wb') as out:
-        out.write(response.content)
-    # 20 bytes is a size of an empty fa.gz
-    if os.path.exists(outpath) and os.path.getsize(outpath) > 20:
-        logging.debug(f"Successful. File saved to {outpath}")
-        return outpath
-    logging.debug(f"Downloaded file {outpath} has zero size. Removing the file.")
-    os.remove(outpath)
-    raise ValueError(f"Downloaded file {outpath} has zero size")
-
-
-def download_from_NCBI_datasets(accession, download_folder):
-    outpath = os.path.join(download_folder, f'{accession}.fa')
-    accession_version = accession if "." in accession else accession + ".1"
-    url = f"https://api.ncbi.nlm.nih.gov/datasets/v2alpha/genome/accession/{accession_version}/download"
-    query = {
-        'include_annotation_type': 'GENOME_FASTA',
-    }
-    response = requests.get(url, params=urllib.parse.urlencode(query))
-    response.raise_for_status()
-
-    content = response.read()
-    tmp_archive = "ncbi_tmp.zip"
-    tmp_archive_path = os.path.join(download_folder, tmp_archive)
-    tmp_path = tmp_archive_path.replace(".zip", "")
-    with open(tmp_archive_path, 'wb') as out:
-        out.write(content)
-    shutil.unpack_archive(tmp_archive_path, tmp_path)
-    subdir_path = os.path.join(download_folder, f"ncbi_tmp/ncbi_dataset/data/{accession_version}/")
-    source_file = [file for file in os.listdir(subdir_path) if file.endswith("_genomic.fna")]
-    source_path = os.path.join(subdir_path, source_file[0]) # assembly_file is a list with one element
-    shutil.move(source_path, outpath)
-    os.remove(tmp_archive_path)
-    shutil.rmtree(tmp_path)
-    if os.path.exists(outpath) and os.path.getsize(outpath) != 0:
-        logging.debug(f"Successful. File saved to {outpath}")
-        return outpath
-    logging.debug(f"Downloaded file {outpath} has zero size. Removing the file.")
-    os.remove(outpath)
-    raise ValueError(f"Downloaded file {outpath} has zero size")
-
-
-@retry(tries=8, delay=10, backoff=3) 
-def download_from_ENA_FTP(accession, outpath):
-    url = get_fasta_url(accession)
-    if not url:
-        logging.debug(f"{accession} URL is empty for accession")
-        return None
-        # raise ValueError(f"URL is empty")
-    logging.debug(f"Download {accession} from ENA FTP using URL {url}")
-    
-    ftp_server = "ftp.ebi.ac.uk"
-    ftp_path = url.replace(ftp_server, "")
-    
-    with FTP(ftp_server) as ftp:
-        ftp.login()
-        with open(outpath, 'wb') as file:
-            ftp.retrbinary(f"RETR {ftp_path}", file.write)
-    # 20 bytes is a size of an empty fa.gz
-    if os.path.exists(outpath) and os.path.getsize(outpath) > 20:
-        logging.debug(f"Successful. File saved to {outpath}")
-        return outpath
-    logging.debug(f"Downloaded file {outpath} has zero size. Removing the file.")
-    os.remove(outpath)
-    return None
-    # raise ValueError(f"Downloaded file {outpath} has zero size")
-
-
 def compare_bin_and_assembly_hashes(acc, mag_hashes, assembly2metadata, download_folder, minchecksum_match):
     confirmed_assemblies = []
     putative_assemblies = {}
@@ -400,6 +295,43 @@ def compare_bin_and_assembly_hashes(acc, mag_hashes, assembly2metadata, download
             if intersection_size >= minchecksum_match:
                 putative_assemblies[assembly] = intersection_size
     return confirmed_assemblies, putative_assemblies
+
+
+def compute_hashes(file_path, write_cache=True, delete_fasta=True, separate_cache_dir=None):
+    hashes = set()
+    cache_path = file_path.replace(".fa.gz", ".fa") + ".hash"
+    if separate_cache_dir:
+        cache_basename = os.path.basename(cache_path)
+        cache_path = os.path.join(separate_cache_dir, cache_basename)
+    
+    if os.path.exists(cache_path):
+        with open(cache_path, "r") as handle:
+            for line in handle:
+                hashes.add(line.strip())
+        return hashes
+
+    if file_path.endswith(".gz"):
+        with gzip.open(file_path, "rt") as handle:
+            for record in SeqIO.parse(handle, "fasta"):
+                hash_object = hashlib.md5(str(record.seq.upper()).encode())
+                hashes.add(hash_object.hexdigest())
+    else:
+        for record in SeqIO.parse(file_path, "fasta"):
+            hash_object = hashlib.md5(str(record.seq.upper()).encode())
+            hashes.add(hash_object.hexdigest())
+
+    if not hashes:
+        raise ValueError("Fasta file does not contain any records")
+
+    if write_cache:
+        with open(cache_path, "w") as handle:
+            for hash in hashes:
+                handle.write(hash + "\n")
+
+    if delete_fasta:
+        os.remove(file_path)
+
+    return hashes
 
 
 def genbank_to_ena_wgsset_accession(acc):
